Refactor incentives e2e tests
