+++ e2e/cypress/features/profile.feature.js
@@ -110,11 +110,6 @@ describe('Profile', () => {
         activity(TODAY, '13:35', '16:45', activities[2], locations[2]),
       ];
 
-      const incentives = {
-        iepLevel: 'Basic',
-        iepDate: '2017-03-08',
-      };
-
       const balances = {
         spends: 123.45,
         cash: 5,
@@ -147,7 +142,6 @@ describe('Profile', () => {
 
       cy.task('stubEvents', todaysEvents);
       cy.task('stubOffenderDetails');
-      cy.task('stubIncentives', incentives);
       cy.task('stubBalancesFor', balances);
       cy.task('stubVisitors', visitors);
       cy.task('stubVisit', visit);
@@ -297,52 +291,6 @@ describe('Profile', () => {
         it("changes to the timetable page when the 'View your full timetable' link is clicked", () => {
           cy.get('[data-test="timetableLink"] > h3 > a').click();
           cy.url().should('include', '/timetable');
-        });
-      });
-    });
-
-    describe('Incentive level (IEP) section', () => {
-      it('displays the expected section heading', () => {
-        cy.get('[data-test="incentive-container"] > h2').should(
-          'contain',
-          'Incentives (IEP)',
-        );
-      });
-
-      describe('Current incentive level panel', () => {
-        it('displays the current incentive level panel', () => {
-          cy.get('[data-test="currentLevel"]').should('exist');
-        });
-
-        it('displays the expected panel heading', () => {
-          cy.get('[data-test="currentLevel"] > h3').should('exist');
-          cy.get('[data-test="currentLevel"] > h3').should(
-            'contain',
-            'My current level:',
-          );
-        });
-
-        it('displays the expected description', () => {
-          cy.get('[data-test="currentLevel"] > p').should('contain', 'Basic');
-        });
-      });
-
-      describe('Incentives link panel', () => {
-        it('displays the incentives link panel', () => {
-          cy.get('[data-test="incentivesLink"]').should('exist');
-        });
-
-        it('displays a link to read more about incentive levels', () => {
-          cy.get('[data-test="incentivesLink"]').should('exist');
-          cy.get('[data-test="incentivesLink"] > span').should(
-            'contain',
-            'Read more about incentives (IEP)',
-          );
-        });
-
-        it("changes to the incentives page when the 'Read more about incentive levels' link is clicked", () => {
-          cy.get('[data-test="incentivesLink"]').click();
-          cy.url().should('include', '/tags/1417');
         });
       });
     });
